Rename to `prefer-bigint-literals`
