+++ docs/rules/prefer-bigint-literals.md
