+++ rules/index.js
@@ -78,7 +78,7 @@ export {default as 'prefer-array-flat'} from './prefer-array-flat.js';
 export {default as 'prefer-array-index-of'} from './prefer-array-index-of.js';
 export {default as 'prefer-array-some'} from './prefer-array-some.js';
 export {default as 'prefer-at'} from './prefer-at.js';
-export {default as 'prefer-bigint-literal'} from './prefer-bigint-literal.js';
+export {default as 'prefer-bigint-literals'} from './prefer-bigint-literals.js';
 export {default as 'prefer-blob-reading-methods'} from './prefer-blob-reading-methods.js';
 export {default as 'prefer-class-fields'} from './prefer-class-fields.js';
 export {default as 'prefer-code-point'} from './prefer-code-point.js';
