+++ test/prefer-bigint-literals.js
