+++ test/snapshots/prefer-bigint-literals.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/prefer-bigint-literal.js`
+# Snapshot report for `test/prefer-bigint-literals.js`
 
-The actual snapshot is saved in `prefer-bigint-literal.js.snap`.
+The actual snapshot is saved in `prefer-bigint-literals.js.snap`.
 
 Generated by [AVA](https://avajs.dev).
